pyproject.toml
[project]
name = "self-representator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "gradio>=6.6.0",
    "llama-index>=0.14.14",
    "llama-index-vector-stores-chroma>=0.5.5",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
